Security-related changes added
Descripción: In this commit, modifications are made to improve the security of the project. Files containing sensitive information such as personal access tokens (PATs) and any other privileged information have been removed. This is part of our ongoing commitment to data protection and the security of information in the repository. Any sensitive information is handled securely and is not included in version control.

#security #privacy #commit #github